requirements.txt
psutil==5.9.6
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
shiboken6==6.6.0
speedtest-cli==2.1.3